+++ NEWS.md
@@ -8,6 +8,7 @@
 * Jobs now write a manifest of files and times to the job tracking database for more thorough completeness tests
 * Added optional low-pass filtering of motion parameters, matching (Gratton)
 * Gracefully adjust motion filtering parameters if they fall above Nyquist at this TR
+* Modify extract ROIs config to avoid input_regex and always generate it internally from postproc stream
 
 # BrainGnomes 0.7-5